Move cmd up one level and add Istio to README
* Istio was not shown in the apps list
* Renamed apps.go to match its command
* Moved cmd out of pkg since it shouldn't be used by outside
consumers

Tested with go build/test.
